Рефакторинг: объединил все компоненты попапа в один
